% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AIC_moveHMM.R
\name{AIC.moveHMM}
\alias{AIC.moveHMM}
\title{AIC}
\usage{
\method{AIC}{moveHMM}(object, ..., k = 2)
}
\arguments{
\item{object}{A \code{moveHMM} object.}

\item{...}{Optional additional \code{moveHMM} objects, to compare AICs of the different models.}

\item{k}{Penalty per parameter. Default : 2 ; for classical AIC.}
}
\value{
The AIC of the model(s) provided.
}
\description{
Akaike information criterion of a moveHMM model.
}
\examples{
m <- ex$m # moveHMM object, as returned by fitHMM
AIC(m)
}

