% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/angleCI.R
\name{angleCI}
\alias{angleCI}
\title{Confidence intervals for angle parameters}
\usage{
angleCI(m)
}
\arguments{
\item{m}{A \code{moveHMM} object}
}
\value{
A list of the following objects :
\item{inf}{Inferior bound of 95\% confidence interval for the parameters of the angle distribution}
\item{sup}{Superior bound of 95\% confidence interval for the parameters of the angle distribution}
}
\description{
Simulation-based computation of confidence intervals for the parameters of the angle distribution.
Used in CI.moveHMM.
}

