% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CI_moveHMM.R
\name{CI}
\alias{CI}
\title{Generic CI method}
\usage{
CI(m, alpha, nbSims)
}
\arguments{
\item{m}{Fitted model}

\item{alpha}{Range of the confidence interval.}

\item{nbSims}{Number of simulations.}
}
\description{
Generic CI method
}

