% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/confIntervals_moveHMM.R
\name{confIntervals}
\alias{confIntervals}
\title{Generic confIntervals method}
\usage{
confIntervals(m)
}
\arguments{
\item{m}{Fitted model}
}
\description{
Generic confIntervals method
}

