% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/decode_moveHMM.R
\name{decode.moveHMM}
\alias{decode.moveHMM}
\title{Decode the model}
\usage{
\method{decode}{moveHMM}(m)
}
\arguments{
\item{m}{\code{moveHMM} object}
}
\value{
List of :
\item{sp}{State probabilities}
\item{res}{Pseudo-residuals}
}
\description{
Decode the model
}
\examples{
m <- ex$mod # moveHMM object (returned by fitHMM)

dec <- decode(m)
dec$sp # state probabilities
dec$res # pseudo-residuals
}

