% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deltaMethod_moveHMM.R
\name{deltaMethod.moveHMM}
\alias{deltaMethod.moveHMM}
\title{Delta method}
\usage{
\method{deltaMethod}{moveHMM}(m)
}
\arguments{
\item{m}{A \code{moveHMM} object.}
}
\value{
A list of :
\code{inf} inferior bound of 95% confidence interval;
\code{sup} superior bound of 95% confidence interval;
both for the parameters of the turning angle distribution.
}
\description{
Computes confidence intervals for the turning angle parameters, using the delta method.
}
\examples{
m <- ex$mod # moveHMM object, as returned by fitHMM

deltaMethod(m)
}

