\name{dexp_rcpp}
\alias{dexp_rcpp}
\title{
  Exponential density function.
}
\description{
  Probability density function of the exponential distribution. Written in C++.
}
\arguments{
  \item{x}{Vector of quantiles}
  \item{rate}{Rate}
  \item{foo}{Unused (compatibility with template)}
}
\value{
  Vector of densities.
}
