\name{dlnorm_rcpp}
\alias{dlnorm_rcpp}
\title{
  Log-normal density function.
}
\description{
  Probability density function of the log-normal distribution. Written in C++.
}
\arguments{
  \item{x}{Vector of quantiles}
  \item{meanlog}{Mean of the distribution on the log scale.}
  \item{sdlog}{Standard deviation of the distribution on the log scale.}
}
\value{
  Vector of densities.
}
