\name{dwrpcauchy_rcpp}
\alias{dwrpcauchy_rcpp}
\title{
  Wrapped-Cauchy density function.
}
\description{
  Probability density function of the wrapped-Cauchy distribution. Written in C++.
}
\arguments{
  \item{x}{Vector of quantiles}
  \item{mu}{Mean}
  \item{rho}{Concentration}
}
\value{
  Vector of densities.
}
