% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitHMM.R
\docType{data}
\name{elk_data}
\alias{elk_data}
\title{Elk data set from Morales et al. (2004, Ecology)}
\usage{
elk_data
}
\description{
It is a data frame with the following columns:
\itemize{
\item \code{ID} Track identifier
\item \code{Easting} Easting coordinate of locations
\item \code{Northing} Northing coordinate of locations
\item \code{dist_water} Distance of elk to water (in metres)
}
}
