
\name{ex}
\docType{data}
\alias{ex}
\title{Example data}
\description{
  This data is generated by the function exGen, and used in the examples and tests of other functions
  to keep them as short as possible.
}
\usage{ex}
\format{
  A list of data (simulated movement data), mod (fitted model), simPar (simulation parameters), and
  par0 (initial parameters used in the fit).
}
