% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitHMM.R
\name{fitHMM}
\alias{fitHMM}
\title{Fit an HMM to the data}
\usage{
fitHMM(nbStates, data, stepPar0, anglePar0, beta0 = NULL, delta0 = NULL,
  formula = ~1, stepDist = c("gamma", "weibull", "lnorm", "exp"),
  angleDist = c("NULL", "vm", "wrpcauchy"), angleMean = NULL,
  zeroInflation = FALSE, verbose = 0)
}
\arguments{
\item{nbStates}{Number of states of the HMM.}

\item{data}{An object moveData.}

\item{stepPar0}{Vector of initial state-dependent step length distribution parameters.}

\item{anglePar0}{Vector of initial state-dependent turning angle distribution parameters.}

\item{beta0}{Initial matrix of regression coefficients for the transition probability matrix.}

\item{delta0}{Initial stationary distribution.}

\item{formula}{Regression formula for the covariates. Default : ~1 (no covariate).}

\item{stepDist}{Name of the distribution of the step length values.}

\item{angleDist}{Name of the distribution of the turning angle values. Defaults to "NULL"
if the turning angles distributions is not estimated.}

\item{angleMean}{Vector of state-dependent turning angles means. It defaults to NULL,
i.e. the means should be estimated.}

\item{zeroInflation}{TRUE if the step length distribution is inflated in zero.}

\item{verbose}{Determines the print level of the optimizer. The default value of 0 means that no
printing occurs, a value of 1 means that the first and last iterations of the optimization are
detailed, and a value of 2 means that each iteration of the optimization is detailed.}
}
\value{
The MLE of the parameters of the model.
}
\description{
Fit an HMM to the data
}
\examples{
data <- example$data
simPar <- example$simPar
par0 <- example$par0

mod <- fitHMM(simPar$nbStates,data,par0$stepPar0,par0$anglePar0,par0$beta0,par0$delta0,
              par0$formula,simPar$stepDist,simPar$angleDist,simPar$angleMean,
              simPar$zeroInflation)
}

