% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPalette}
\alias{getPalette}
\title{Discrete colour palette for states}
\usage{
getPalette(nbStates)
}
\arguments{
\item{nbStates}{Number of states}
}
\value{
Vector of colours, of length nbStates.
}
\description{
Discrete colour palette for states
}
