% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitHMM.R
\docType{data}
\name{haggis_data}
\alias{haggis_data}
\title{Wild haggis data set from Michelot et al. (2016, Methods Eco Evol)}
\usage{
haggis_data
}
\description{
Data frame of the first three tracks from Michelot et al. (2016),
 with columns:
\itemize{
\item \code{ID} Track identifier
\item \code{x} Easting coordinate of locations
\item \code{y} Northing coordinate of locations
\item \code{slope} Terrain slope (in degrees)
\item \code{temp} Air temperature (in degrees Celsius)
}
}
