% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveHMM.R
\name{is.moveHMM}
\alias{is.moveHMM}
\title{Is moveHMM}
\usage{
is.moveHMM(x)
}
\arguments{
\item{x}{An R object}
}
\value{
\code{TRUE} if \code{x} is of class \code{\link{moveHMM}}, \code{FALSE} otherwise.
}
\description{
Check that an object is of class \code{\link{moveHMM}}. Used in \code{\link{CI}},
\code{\link{plotPR}}, \code{\link{plotStates}}, \code{\link{pseudoRes}}, \code{\link{stateProbs}},
and \code{\link{viterbi}}.
}
