% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lalpha_moveHMM.R
\name{lalpha}
\alias{lalpha}
\title{Forward probabilities}
\usage{
lalpha(m)
}
\arguments{
\item{m}{A moveHMM object.}
}
\value{
The matrix of forward probabilities.
}
\description{
Forward probabilities
}

