% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lbeta_moveHMM.R
\name{lbeta}
\alias{lbeta}
\title{Backward probabilities}
\usage{
lbeta(m)
}
\arguments{
\item{m}{A moveHMM object.}
}
\value{
The matrix of backward probabilities.
}
\description{
Backward probabilities
}
\examples{
m <- example$mod # moveHMM object (returned by fitHMM)

lb <- lbeta(m)
}

