% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logAlpha_moveHMM.R
\name{logAlpha.moveHMM}
\alias{logAlpha.moveHMM}
\title{Forward log-probabilities}
\usage{
\method{logAlpha}{moveHMM}(m)
}
\arguments{
\item{m}{A moveHMM object.}
}
\value{
The matrix of forward log-probabilities.
}
\description{
Forward log-probabilities
}
\examples{
m <- example$mod # moveHMM object (returned by fitHMM)

la <- logAlpha(m)
}

