% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logBeta.R
\name{logBeta}
\alias{logBeta}
\title{Backward log-probabilities}
\usage{
logBeta(m)
}
\arguments{
\item{m}{A moveHMM object.}
}
\value{
The matrix of backward log-probabilities.
}
\description{
Backward log-probabilities
}
\examples{
m <- ex$mod # moveHMM object (returned by fitHMM)

lb <- logBeta(m)
}

