% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/moveData.R
\name{moveData}
\alias{moveData}
\title{Constructor of moveData objects}
\usage{
moveData(data)
}
\arguments{
\item{data}{A dataframe containing the IDs, step lengths, turning angles (if any),
and covariates (if any), as well as the x/y values (either cartesian coordinates
or longitude/latitude)}
}
\value{
An object moveData.
}
\description{
Constructor of moveData objects
}

