% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/n2w.R
\name{n2w}
\alias{n2w}
\title{Scaling function : natural to working parameters.}
\usage{
n2w(par, bounds, beta, delta, nbStates, estAngleMean)
}
\arguments{
\item{par}{Vector of state-dependent distributions parameters.}

\item{bounds}{Matrix with 2 columns and as many rows as there are elements in par. Each row
contains the lower and upper bound for the correponding parameter.}

\item{beta}{Matrix of regression coefficients for the transition probability matrix.}

\item{delta}{Stationary distribution.}

\item{nbStates}{The number of states of the HMM.}

\item{estAngleMean}{TRUE if the angle mean is estimated, FALSE otherwise.}
}
\value{
A vector of unconstrained parameters.
}
\description{
Scales each parameter from its natural interval to the set of real numbers, to allow for
unconstrained optimization. Used during the optimization of the log-likelihood.
}
\examples{
nbStates <- 3
par <- c(0.001,0.999,0.5,0.001,1500.3,7.1)
bounds <- matrix(c(0,1,0,1,0,1,
                   0,Inf,0,Inf,0,Inf),
                 byrow=TRUE,ncol=2)
beta <- matrix(rnorm(18),ncol=6,nrow=3)
delta <- c(0.6,0.3,0.1)
print(n2w(par,bounds,beta,delta,nbStates))
}

