% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nLogLike.R
\name{nLogLike}
\alias{nLogLike}
\title{Negative log-likelihood function}
\usage{
nLogLike(wpar, nbStates, bounds, parSize, data, stepDist = c("gamma",
  "weibull", "exp"), angleDist = c("NULL", "vm", "wrpcauchy"),
  angleMean = NULL, zeroInflation = FALSE)
}
\arguments{
\item{wpar}{Vector of working parameters.}

\item{nbStates}{Number of states of the HMM.}

\item{bounds}{Matrix with 2 columns and as many rows as there are elements in wpar. Each row
contains the lower and upper bound for the correponding parameter.}

\item{parSize}{Vector of two values : c(number of parameters of the step length distribution,
number of parameters of the turning angle distribution).}

\item{data}{An object moveData.}

\item{stepDist}{Name of the distribution of the step length values.}

\item{angleDist}{Name of the distribution of the turning angle values. Defaults to "NULL"
if the turning angles distributions is not estimated.}

\item{angleMean}{Vector of means of turning angles if not estimated (one for each state).
Defaults to NULL.}

\item{zeroInflation}{TRUE if the step length distribution is inflated in zero.}
}
\value{
The negative log-likelihood of wpar given data.
}
\description{
Negative log-likelihood function
}
\examples{
data <- example$data

nbStates <- 2
nbCovs <- 2
stepDist <- "gamma"
angleDist <- "vm"

bounds <- parDef(stepDist,angleDist,nbStates,FALSE,FALSE)$bounds
parSize <- parDef(stepDist,angleDist,nbStates,FALSE,FALSE)$parSize

mu0 <- c(20,80)
sigma0 <- c(20,40)
kappa0 <- c(1,1)
beta0 <- matrix(c(rep(-1.5,nbStates*(nbStates-1)),rep(0,nbStates*(nbStates-1)*nbCovs)),
                nrow=nbCovs+1,byrow=TRUE)
delta0 <- c(1,1)/2
par0 <- c(mu0,sigma0,kappa0)
wpar <- n2w(par0,bounds,beta0,delta0,nbStates)
angleMean <- c(pi,0)

l <- nLogLike(wpar,nbStates,bounds,parSize,data,stepDist,angleDist,angleMean,FALSE)
}

