% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parDef.R
\name{parDef}
\alias{parDef}
\title{Parameters definition}
\usage{
parDef(stepDist = c("gamma", "weibull", "exp"), angleDist = c("vm",
  "wrpcauchy"), nbStates, estAngleMean)
}
\arguments{
\item{stepDist}{Name of the distribution of the step length values.}

\item{angleDist}{Name of the distribution of the turning angle values.}

\item{nbStates}{Number of states of the HMM.}

\item{estAngleMean}{TRUE if the mean of the turning angles distribution is estimated,
FALSE otherwise.}
}
\value{
A list of two elements : parSize (= c(number of parameters of the step length
distribution,number of parameters of the turning angle distribution)) and bounds
(matrix with 2 columns and sum(parSize) rows. Each row contains the lower and upper
bound for the correponding parameter).
}
\description{
Parameters definition
}

