% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_moveHMM.R
\name{plot.moveHMM}
\alias{plot.moveHMM}
\title{Plot moveHMM}
\usage{
\method{plot}{moveHMM}(x, ask = TRUE, animals = NULL, breaks = "Sturges",
  hist.ylim = NULL, ...)
}
\arguments{
\item{x}{Object moveHMM}

\item{ask}{If TRUE, the execution pauses between each plot.}

\item{animals}{Vector of indices of animals for which information will be plotted.
Default : NULL ; all animals are plotted.}

\item{breaks}{Histogram parameter. See hist documentation.}

\item{hist.ylim}{Parameter ylim for the step length histograms. See hist documentation.
Default : NULL ; the function sets default values.}

\item{...}{Currently unused. For compatibility with generic method.}
}
\description{
Plot moveHMM
}
\examples{
m <- ex$mod # moveHMM object, as returned by fitHMM

plot(m,ask=TRUE,animals=1,breaks=20)
}

