% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotPR.R
\name{plotPR}
\alias{plotPR}
\title{Plot pseudo-residuals}
\usage{
plotPR(m)
}
\arguments{
\item{m}{A \code{\link{moveHMM}} object}
}
\description{
Plots qq-plots and time series of the pseudo-residuals
}
\examples{
m <- ex$m # moveHMM object (as returned by \\code{\\link{fitHMM}})

plotPR(m)
}

