% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotStates.R
\name{plotStates.moveHMM}
\alias{plotStates.moveHMM}
\title{Plot states}
\usage{
\method{plotStates}{moveHMM}(m, animals = NULL)
}
\arguments{
\item{m}{A \code{\link{moveHMM}} object}

\item{animals}{Vector of indices or IDs of animals for which states will be plotted.}
}
\description{
Plot the states and states probabilities.
}
\examples{
m <- ex$m # moveHMM object, as returned by fitHMM

# plot states for first and second animals
plotStates(m,animals=c(1,2))
}

