% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotStates.R
\name{plotStates}
\alias{plotStates}
\title{Generic plotStates method}
\usage{
plotStates(m, animals)
}
\arguments{
\item{m}{Fitted model}

\item{animals}{Animals to include}
}
\description{
Generic plotStates method
}

