% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStationary.R
\name{plotStationary}
\alias{plotStationary}
\title{Plot stationary state probabilities}
\usage{
plotStationary(m, col = NULL, plotCI = FALSE)
}
\arguments{
\item{m}{An object \code{moveHMM}}

\item{col}{Vector or colors for the states (one color per state).}

\item{plotCI}{Logical. Should 95\% confidence intervals be plotted? (Default: FALSE)}
}
\description{
Plot stationary state probabilities
}
