% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print_moveHMM.R
\name{print.moveHMM}
\alias{print.moveHMM}
\title{Print moveHMM}
\usage{
\method{print}{moveHMM}(x, ...)
}
\arguments{
\item{x}{A moveHMM object.}

\item{...}{Currently unused. For compatibility with generic method.}
}
\description{
Print moveHMM
}
\examples{
m <- ex$mod # moveHMM object (returned by fitHMM)

print(m)
}

