% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pseudoRes_moveHMM.R
\name{pseudoRes.moveHMM}
\alias{pseudoRes.moveHMM}
\title{Pseudo-residuals}
\usage{
\method{pseudoRes}{moveHMM}(m)
}
\arguments{
\item{m}{A \code{moveHMM} object.}
}
\value{
A list of :
\item{stepRes}{The pseudo-residuals for the step lengths}
\item{angleRes}{The pseudo-residuals for the turning angles}
}
\description{
Pseudo-residuals
}
\examples{
m <- ex$mod # moveHMM object (returned by fitHMM)
res <- pseudoRes(m)
qqnorm(res$stepRes)
qqnorm(res$angleRes)
}

