% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pseudoRes.R
\name{pseudoRes}
\alias{pseudoRes}
\title{Pseudo-residuals}
\usage{
pseudoRes(m)
}
\arguments{
\item{m}{A \code{moveHMM} object.}
}
\value{
A list of:
\item{stepRes}{The pseudo-residuals for the step lengths}
\item{angleRes}{The pseudo-residuals for the turning angles}
}
\description{
Pseudo-residuals
}
\details{
If some turning angles in the data are equal to pi, the corresponding pseudo-residuals
will not be included. Indeed, given that the turning angles are defined on ]-pi,pi], an angle of pi
results in a pseudo-residual on +Inf (check Section 6.2 of reference for more information on the
computation of pseudo-residuals).
}
\examples{
m <- ex$m # moveHMM object (returned by fitHMM)
res <- pseudoRes(m)
qqnorm(res$stepRes)
qqnorm(res$angleRes)
}
\references{
Zucchini, W. and MacDonald, I.L. 2009.
Hidden Markov Models for Time Series: An Introduction Using R.
Chapman & Hall (London).
}

