% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{Simulation tool}
\usage{
simData(nbAnimals, nbStates, stepDist = c("gamma", "weibull", "exp"),
  angleDist = c("vm", "wrpcauchy"), stepPar, anglePar, zeroInflation = 0,
  nbCov = 0)
}
\arguments{
\item{nbAnimals}{Number of observed individuals to simulate.}

\item{nbStates}{Number of behavioural states to simulate.}

\item{stepDist}{Name of the distribution from which to draw the step length values.}

\item{angleDist}{Name of the distribution from which to draw the turning angle values.}

\item{stepPar}{Parameters of the step length distribution. Must be provided in a
matrix with one row for each parameter (in the order expected by the pdf of stepDist),
and one column for each state.}

\item{anglePar}{Parameters of the turning angle distribution. Must be provided in a
matrix with one row for each parameter (in the order expected by the pdf of angleDist),
and one column for each state.}

\item{zeroInflation}{Probability of step length being zero (0 by default).}

\item{nbCov}{Number of covariates to simulate (0 by default).}
}
\value{
An object moveData
}
\description{
Simulation tool
}
\examples{
stepPar <- matrix(c(1,1,10,5),nrow=2) # mean1, sd1, mean2, sd2
anglePar <- matrix(c(0,0.5,pi,2),nrow=2) # mean1, k1, mean2, k2
stepDist <- "gamma"
angleDist <- "vm"
data <- simData(5,2,stepDist,angleDist,stepPar,anglePar,0.2,2)

stepPar <- matrix(c(1,1,10,5),nrow=2) # mean1, sd1, mean2, sd2
anglePar <- matrix(c(0,0.5,pi,0.7),nrow=2) # mean1, k1, mean2, k2
stepDist <- "weibull"
angleDist <- "wrpcauchy"
data <- simData(5,2,stepDist,angleDist,stepPar,anglePar)
}

