% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stateProbs_moveHMM.R
\name{stateProbs}
\alias{stateProbs}
\title{State probabilities}
\usage{
stateProbs(m)
}
\arguments{
\item{m}{A moveHMM object.}
}
\value{
The matrix of state probabilities.
}
\description{
State probabilities
}
\examples{
m <- example$mod # moveHMM object (returned by fitHMM)

sp <- stateProbs(m)
}

