% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationary.R
\name{stationary}
\alias{stationary}
\title{Stationary state probabilities}
\usage{
stationary(m, covMat)
}
\arguments{
\item{m}{Fitted model (as output by \code{\link{fitHMM}}).}

\item{covMat}{Design matrix of covariates.}
}
\value{
Matrix of stationary state probabilities. Each row corresponds to
a row of covMat, and each column corresponds to a state.
}
\description{
Calculates the stationary probabilities of each state, for given
covariate values.
}
