% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_moveData.R
\name{summary.moveData}
\alias{summary.moveData}
\title{Summary \code{moveData}}
\usage{
\method{summary}{moveData}(object, ...)
}
\arguments{
\item{object}{A \code{moveData} object.}

\item{...}{Currently unused. For compatibility with generic method.}
}
\description{
Summary \code{moveData}
}
\examples{
# Use example data set as output by prepData
summary(example$data)

}
