% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trMatrix.R
\name{trMatrix}
\alias{trMatrix}
\title{Generate the transition probability matrix}
\usage{
trMatrix(nbStates, beta, covs)
}
\arguments{
\item{nbStates}{Number of states of the HMM.}

\item{beta}{Matrix of regression parameters. Number of rows : number of covariates + 1 ;
number of columns : number of non-diagonal elements in the t.p.m. (i.e. nbStates*(nbStates-1)).}

\item{covs}{Design matrix of covariates.}
}
\value{
A three-dimensional array gamma, such that gamma[,,t] is the transition probability
matrix corresponding to observation t.
}
\description{
Used in the computation of the log-likelihood.
}

