% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/w2n.R
\name{w2n}
\alias{w2n}
\title{Scaling function : working to natural parameters}
\usage{
w2n(wpar, bounds, nbStates, nbCovs)
}
\arguments{
\item{wpar}{Vector of state-dependent distributions  unconstrained parameters.}

\item{bounds}{Matrix with 2 columns and as many rows as there are elements in par. Each row
contains the lower and upper bound for the correponding parameter.}

\item{nbStates}{The number of states of the HMM.}

\item{nbCovs}{The number of covariates.}
}
\value{
A list containing a vector of natural (constrained) parameters, as well as delta
and beta.
}
\description{
Scales each parameter from the set of real numbers, back to its natural interval.
}
\examples{
nbStates <- 3
nbCovs <- 2
par <- c(0.001,0.999,0.5,0.001,1500.3,7.1)
bounds <- matrix(c(0,1,0,1,0,1,
                   0,Inf,0,Inf,0,Inf),
                 byrow=T,ncol=2)
beta <- matrix(rnorm(18),ncol=6,nrow=3)
delta <- c(0.6,0.3,0.1)
wpar <- n2w(par,bounds,beta,delta,nbStates)
print(w2n(wpar,bounds,nbStates,nbCovs))
}

