% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC_moveHMM.R
\name{AIC.moveHMM}
\alias{AIC.moveHMM}
\title{AIC}
\usage{
\method{AIC}{moveHMM}(object, ..., k = 2)
}
\arguments{
\item{object}{A \code{moveHMM} object.}

\item{...}{Optional additional \code{moveHMM} objects, to compare AICs of the different models.}

\item{k}{Penalty per parameter. Default: 2; for classical AIC.}
}
\value{
The AIC of the model(s) provided. If several models are provided, the AICs are output
in ascending order.
}
\description{
Akaike information criterion of a moveHMM model.
}
\examples{
# m is a moveHMM object (as returned by fitHMM), automatically loaded with the package
m <- example$m
AIC(m)

}
