% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allProbs.R
\name{allProbs}
\alias{allProbs}
\title{Matrix of all probabilities}
\usage{
allProbs(data, nbStates, stepDist = c("gamma", "weibull", "exp"),
  angleDist = c("NULL", "vm", "wrpcauchy"), stepPar, anglePar = NULL,
  zeroInflation = FALSE)
}
\arguments{
\item{data}{Data relative to one observed individual, as a list. Required fields :
step, angle.}

\item{nbStates}{Number of states of the HMM.}

\item{stepDist}{Name of the distribution of the step length values.}

\item{angleDist}{Name of the distribution of the turning angle values. Defaults to "NULL"
if the turning angles distributions is not estimated.}

\item{stepPar}{Parameters of the step length distribution. Must be provided in a
matrix with one row for each parameter (in the order expected by the pdf of stepDist),
and one column for each state.}

\item{anglePar}{Parameters of the turning angle distribution. Must be provided in a
matrix with one row for each parameter (in the order expected by the pdf of angleDist),
and one column for each state. Defaults to NULL if the turning angles distributions
is not estimated.}

\item{zeroInflation}{TRUE if the step length distribution is inflated in zero.}
}
\value{
Matrix of all probabilities.
}
\description{
Used in the computation of the log-likelihood.
}
\examples{
stepPar <- matrix(c(1,1,10,5),nrow=2) # mean1, sd1, mean2, sd2
anglePar <- matrix(c(0,0.5,pi,2),nrow=2) # mean1, k1, mean2, k2
stepDist <- "gamma"
angleDist <- "vm"
data <- simData(5,2,stepDist,angleDist,stepPar,anglePar,0.2,2)
P <- allProbs(data[[1]],2,stepDist,angleDist,stepPar,anglePar)
}

