% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allProbs.R
\name{allProbs}
\alias{allProbs}
\title{Matrix of all probabilities}
\usage{
allProbs(data, nbStates, stepFun = c("gamma", "weibull", "exp"),
  angleFun = c("vm", "wrpcauchy"), stepPar, anglePar)
}
\arguments{
\item{data}{Data relative to one observed individual, as a list. Required fields :
step, angle.}

\item{nbStates}{Number of states of the HMM.}

\item{stepFun}{Character string, among "gamma", "weibull", and "exp".
Name of the distribution of the step length values.}

\item{angleFun}{Character string, among "vm", and "wrpcauchy".
Name of the distribution of the turning angle values.}

\item{stepPar}{Parameters of the step length distribution. Must be provided in a
matrix with one row for each parameter (in the order expected by the function stepFun),
and one column for each state.}

\item{anglePar}{Parameters of the turning angle distribution. Must be provided in a
matrix with one row for each parameter (in the order expected by the function angleFun),
and one column for each state.}
}
\value{
Matrix of all probabilities.
}
\description{
Used in the computation of the log-likelihood.
}
\examples{
stepPar <- matrix(c(1,1,10,5),nrow=2) # mean1, sd1, mean2, sd2
anglePar <- matrix(c(0,0.5,pi,2),nrow=2) # mean1, k1, mean2, k2
stepFun <- "gamma"
angleFun <- "vm"
data <- simData(5,2,stepFun,angleFun,stepPar,anglePar,0.2,2)
P <- allProbs(data[[1]],2,stepFun,angleFun,stepPar,anglePar)
}

