% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/angleCI.R
\name{angleCI}
\alias{angleCI}
\title{Confidence intervals for angle parameters}
\usage{
angleCI(m, alpha)
}
\arguments{
\item{m}{A \code{moveHMM} object}

\item{alpha}{Range of the confidence intervals. Default : 0.95 (i.e. 95% CIs).}
}
\value{
A list of the following objects :
\item{inf}{Inferior bound of the confidence interval for the parameters of the angle distribution}
\item{sup}{Superior bound of the confidence interval for the parameters of the angle distribution}
}
\description{
Simulation-based computation of confidence intervals for the parameters of the angle distribution.
Used in \code{\link{CI.moveHMM}}.
}

