% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CI_moveHMM.R
\name{CI.moveHMM}
\alias{CI.moveHMM}
\title{Confidence intervals}
\usage{
\method{CI}{moveHMM}(m, alpha = 0.95, nbSims = 10^6)
}
\arguments{
\item{m}{A \code{moveHMM} object}

\item{alpha}{Range of the confidence intervals. Default : 0.95 (i.e. 95\% CIs).}

\item{nbSims}{Number of simulations in the computation of the CIs for the angle parameters.
Default : 10^6.}
}
\value{
A list of the following objects :
\item{lower}{Lower bound of the confidence interval for the parameters of the step lengths
distribution, for the parameters of the turning angle distribution, and for the transition
probabilities regression parameters}
\item{upper}{Upper bound of the confidence interval for the parameters of the step lengths
distribution, for the parameters of the turning angle distribution, and for the transition
probabilities regression parameters}
}
\description{
Computes the confidence intervals of the step length and turning angle parameters,
as well as for the transition probabilities regression parameters.
}
\examples{
m <- ex$m # moveHMM object, as returned by fitHMM

CI(m)
}

