% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CI_moveHMM.R
\name{CI.moveHMM}
\alias{CI.moveHMM}
\title{Confidence intervals}
\usage{
\method{CI}{moveHMM}(m)
}
\arguments{
\item{m}{A \code{moveHMM} object}
}
\value{
A list of the following objects :
\item{inf}{Inferior bound of 95% confidence interval for the parameters of the step lengths
distribution, for the parameters of the turning angle distribution, and for the transition
probabilities regression parameters}
\item{sup}{Superior bound of 95% confidence interval for the parameters of the step lengths
distribution, for the parameters of the turning angle distribution, and for the transition
probabilities regression parameters}
}
\description{
Computes the confidence intervals of the step length and turning angle parameters,
as well as for the transition probabilities regression parameters.
}
\examples{
m <- ex$mod # moveHMM object, as returned by fitHMM

CI(m)
}

