% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitHMM.R
\name{fitHMM}
\alias{fitHMM}
\title{Fit an HMM to the data}
\usage{
fitHMM(nbStates, data, par0, beta0, delta0, stepDist = c("gamma", "weibull",
  "exp"), angleDist = c("NULL", "vm", "wrpcauchy"), angleMean = NULL)
}
\arguments{
\item{nbStates}{Number of states of the HMM.}

\item{data}{An object moveData.}

\item{par0}{Vector of initial state-dependent distributions parameters.}

\item{beta0}{Initial matrix of regression coefficients for the transition probability matrix.}

\item{delta0}{Initial stationary distribution.}

\item{stepDist}{Name of the distribution of the step length values.}

\item{angleDist}{Name of the distribution of the turning angle values. Defaults to "NULL"
if the turning angles distributions is not estimated.}

\item{angleMean}{If the mean of the turning angle distribution should not be estimated,
angleMean is the vector of state-dependent means. It defaults to NULL, i.e. the means
should be estimated.}
}
\value{
The MLE of the parameters of the model.
}
\description{
Fit an HMM to the data
}
\examples{
nbStates <- 2
stepPar <- matrix(c(15,50,
                    10,20),
                  byrow=TRUE,ncol=2)
anglePar <- matrix(c(pi,0,
                     0.7,2),
                   byrow=TRUE,ncol=2)
data <- simData(2,nbStates,"gamma","vm",stepPar,anglePar,nbCov=2)

mu0 <- c(20,80)
sigma0 <- c(20,40)
kappa0 <- c(1,1)
par0 <- c(mu0,sigma0,kappa0)

nbCovs <- ncol(data[[1]]$covs)
beta0 <- matrix(c(rep(-1.5,nbStates*(nbStates-1)),rep(0,nbStates*(nbStates-1)*nbCovs)),
                nrow=nbCovs+1,byrow=TRUE)
delta0 <- c(1,1)/2

mod <- fitHMM(nbStates,data,par0,beta0,delta0,"gamma","vm",c(pi,0))
}

