% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logAlpha.R
\name{logAlpha}
\alias{logAlpha}
\title{Forward log-probabilities}
\usage{
logAlpha(m)
}
\arguments{
\item{m}{A \code{moveHMM} object.}
}
\value{
The matrix of forward log-probabilities.
}
\description{
Used in \code{\link{stateProbs}} and \code{\link{pseudoRes}}.
}
\examples{
m <- ex$mod # moveHMM object (returned by fitHMM)

la <- logAlpha(m)
}

