% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/moveHMM.R
\name{moveHMM}
\alias{moveHMM}
\title{Constructor of moveHMM objects}
\usage{
moveHMM(m)
}
\arguments{
\item{mod}{A list of attributes of the fitted model. Mandatory slots : data, states, mle,
stepDist, angleDist, mod (result of call to optimizer),
Optional slots : angleMean (if not estimated)}
}
\value{
An object moveHMM.
}
\description{
Constructor of moveHMM objects
}

