% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nLogLike.R
\name{nLogLike}
\alias{nLogLike}
\title{Negative log-likelihood function}
\usage{
nLogLike(wpar, nbStates, bounds, parSize, data, stepDist = c("gamma",
  "weibull", "lnorm", "exp"), angleDist = c("vm", "wrpcauchy", "none"),
  angleMean = NULL, zeroInflation = FALSE)
}
\arguments{
\item{wpar}{Vector of working parameters.}

\item{nbStates}{Number of states of the HMM.}

\item{bounds}{Matrix with 2 columns and as many rows as there are elements in wpar. Each row
contains the lower and upper bound for the correponding parameter.}

\item{parSize}{Vector of two values : c(number of parameters of the step length distribution,
number of parameters of the turning angle distribution).}

\item{data}{An object moveData.}

\item{stepDist}{Name of the distribution of the step lengths.
Supported distributions are : gamma, weibull, lnorm, exp.}

\item{angleDist}{Name of the distribution of the turning angles.
Supported distributions are : vm, wrpcauchy. Set to "none" if the angle distribution should
not be estimated.}

\item{angleMean}{Vector of means of turning angles if not estimated (one for each state).
Defaults to NULL.}

\item{zeroInflation}{TRUE if the step length distribution is inflated in zero.}
}
\value{
The negative log-likelihood of wpar given data.
}
\description{
Negative log-likelihood function
}
\examples{
data <- example$data
simPar <- example$simPar
par0 <- example$par0

estAngleMean <- is.null(simPar$angleMean)
bounds <- parDef(simPar$stepDist,simPar$angleDist,simPar$nbStates,
                 estAngleMean,simPar$zeroInflation)$bounds
parSize <- parDef(simPar$stepDist,simPar$angleDist,simPar$nbStates,
                  estAngleMean,simPar$zeroInflation)$parSize

par <- c(par0$stepPar0,simPar$angleMean,par0$anglePar0)
wpar <- n2w(par,bounds,par0$beta0,par0$delta0,simPar$nbStates)

l <- nLogLike(wpar,simPar$nbStates,bounds,parSize,data,simPar$stepDist,simPar$angleDist,
              simPar$angleMean,simPar$zeroInflation)
}

