% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parDef.R
\name{parDef}
\alias{parDef}
\title{Parameters definition}
\usage{
parDef(stepDist, angleDist, nbStates, estAngleMean, zeroInflation)
}
\arguments{
\item{stepDist}{Name of the distribution of the step length values.}

\item{angleDist}{Name of the distribution of the turning angle values. Defaults to "none"
if the turning angles distributions is not estimated.}

\item{nbStates}{Number of states of the HMM.}

\item{estAngleMean}{TRUE if the mean of the turning angles distribution is estimated,
FALSE otherwise.}

\item{zeroInflation}{TRUE if the step length distribution is inflated in zero.}
}
\value{
A list of three elements : parSize (= c(number of parameters of the step length
distribution,number of parameters of the turning angle distribution)), bounds
(matrix with 2 columns and sum(parSize) rows - each row contains the lower and upper
bound for the correponding parameter), and parNames (names of parameters).
}
\description{
Parameters definition
}

