% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_moveData.R
\name{plot.moveData}
\alias{plot.moveData}
\title{Plot function for a moveData object}
\usage{
\method{plot}{moveData}(x, compact = FALSE, breaks = "Sturges", ...)
}
\arguments{
\item{x}{An object moveData}

\item{compact}{TRUE for a compact plot (all individuals at once), FALSE otherwise (default -- one
individual at a time)}

\item{breaks}{Histogram parameter. See hist documentation.}

\item{...}{Currently unused. For compatibility with generic method.}
}
\description{
Plot function for a moveData object
}

