% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_moveData.R
\name{plot.moveData}
\alias{plot.moveData}
\title{Plot function for a moveData object}
\usage{
plot.moveData(data, compact = FALSE)
}
\arguments{
\item{data}{An object moveData}

\item{compact}{TRUE for a compact plot (all individuals at once), FALSE otherwise (default -- one
individual at a time)}
}
\description{
Plot function for a moveData object
}

