% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_moveData.R
\name{plot.moveData}
\alias{plot.moveData}
\title{Plot \code{moveData}}
\usage{
\method{plot}{moveData}(x, compact = FALSE, ask = TRUE,
  breaks = "Sturges", ...)
}
\arguments{
\item{x}{An object \code{moveData}}

\item{compact}{\code{TRUE} for a compact plot (all individuals at once), \code{FALSE} otherwise
(default -- one individual at a time).}

\item{ask}{If \code{TRUE}, the execution pauses between each plot.}

\item{breaks}{Histogram parameter. See \code{hist} documentation.}

\item{...}{Currently unused. For compatibility with generic method.}
}
\description{
Plot \code{moveData}
}
\examples{
data <- ex$data # moveData object, as returned by prepData or simData

plot(data,compact=TRUE,breaks=20,ask=FALSE)
}

