% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_moveHMM.R
\name{plot.moveHMM}
\alias{plot.moveHMM}
\title{Plot \code{moveHMM}}
\usage{
\method{plot}{moveHMM}(x, animals = NULL, ask = TRUE, breaks = "Sturges",
  hist.ylim = NULL, compactHist = FALSE, ...)
}
\arguments{
\item{x}{Object \code{moveHMM}}

\item{animals}{Vector of indices or IDs of animals for which information will be plotted.
Default : \code{NULL} ; all animals are plotted.}

\item{ask}{If \code{TRUE}, the execution pauses between each plot.}

\item{breaks}{Histogram parameter. See \code{hist} documentation.}

\item{hist.ylim}{Parameter \code{ylim} for the step length histograms. See \code{hist} documentation.
Default : \code{NULL} ; the function sets default values.}

\item{compactHist}{If \code{TRUE}, the function only plots histograms of all observations for steps
and angles, with the fitted densities (no map, and no individual-specific plot). Default : \code{FALSE}.}

\item{...}{Currently unused. For compatibility with generic method.}
}
\description{
Plot \code{moveHMM}
}
\examples{
m <- ex$mod # moveHMM object, as returned by fitHMM

plot(m,ask=TRUE,animals=1,breaks=20)
}

