% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_moveHMM.R
\name{plot.moveHMM}
\alias{plot.moveHMM}
\title{Plot moveHMM}
\usage{
\method{plot}{moveHMM}(x, ...)
}
\arguments{
\item{x}{Object moveHMM}

\item{...}{Currently unused. For compatibility with generic method.}
}
\description{
Plot moveHMM
}

