% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotStates.R
\name{plotStates}
\alias{plotStates}
\title{Plot states}
\usage{
plotStates(m, animals = NULL, ask = TRUE)
}
\arguments{
\item{m}{A \code{\link{moveHMM}} object}

\item{animals}{Vector of indices or IDs of animals for which states will be plotted.}

\item{ask}{If \code{TRUE}, the execution pauses between each plot.}
}
\description{
Plot the states and states probabilities.
}
\examples{
# m is a moveHMM object (as returned by fitHMM), automatically loaded with the package
m <- example$m

# plot states for first and second animals
plotStates(m,animals=c(1,2))
}

