% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepData.R
\name{prepData}
\alias{prepData}
\title{Preprocessing of the tracking data}
\usage{
prepData(trackData, type = c("GCD", "euclidean"), coordNames = c("x", "y"))
}
\arguments{
\item{trackData}{A dataframe of the tracking data, including at least coordinates
(either longitude/latitude values or cartesian coordinates), and optionnaly a field "ID"
(identifiers for the observed individuals). Additionnal fields are considered as covariates.
Note that, if the names of the coordinates are not "x" and "y", the \code{coordNames} argument
should specified.}

\item{type}{'GCD' if longitude/latitude provided (default), 'euclidean' otherwise.}

\item{coordNames}{Names of the columns of coordinates in the data frame. Default : c("x","y").}
}
\value{
An object moveData, i.e. a dataframe of ID, step lengths,
turning angles and covariates values (if any).
}
\description{
Preprocessing of the tracking data
}
\examples{
coord1 <- c(1,2,3,4,5,6,7,8,9,10)
coord2 <- c(1,1,1,2,2,2,1,1,1,2)
trackData <- data.frame(coord1=coord1,coord2=coord2)
d <- prepData(trackData,'euclidean',coordNames=c("coord1","coord2"))
}

