% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print_moveHMM.R
\name{print.moveHMM}
\alias{print.moveHMM}
\title{Print moveHMM}
\usage{
\method{print}{moveHMM}(m)
}
\arguments{
\item{m}{A moveHMM object.}
}
\description{
Print moveHMM
}
\examples{
m <- example$mod # moveHMM object (returned by fitHMM)

print(m)
}

