% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pseudoRes_moveHMM.R
\name{pseudoRes.moveHMM}
\alias{pseudoRes.moveHMM}
\title{Pseudo-residuals}
\usage{
\method{pseudoRes}{moveHMM}(m)
}
\arguments{
\item{m}{A moveHMM object.}
}
\value{
The pseudo-residuals for the step lengths, stepRes, and the pseudo-residuals for
the turning angles, angleRes.
}
\description{
Pseudo-residuals
}
\examples{
m <- example$mod # moveHMM object (returned by fitHMM)
res <- pseudoRes(m)
qqnorm(res$stepRes)
qqnorm(res$angleRes)
}

