% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{Simulation tool}
\usage{
simData(nbAnimals, nbStates, stepDist = c("gamma", "weibull", "lnorm", "exp"),
  angleDist = c("vm", "wrpcauchy", "none"), stepPar, anglePar = NULL,
  beta = NULL, covs = NULL, nbCovs = 0, zeroInflation = FALSE,
  obsPerAnimal = c(500, 1500))
}
\arguments{
\item{nbAnimals}{Number of observed individuals to simulate.}

\item{nbStates}{Number of behavioural states to simulate.}

\item{stepDist}{Name of the distribution of the step lengths (as a character string).
Supported distributions are : gamma, weibull, lnorm, exp. Default : gamma.}

\item{angleDist}{Name of the distribution of the turning angles (as a character string).
Supported distributions are : vm, wrpcauchy. Set to \code{"none"} if the angle distribution should
not be estimated. Default : vm.}

\item{stepPar}{Parameters of the step length distribution.}

\item{anglePar}{Parameters of the turning angle distribution.}

\item{beta}{Matrix of regression parameters for the transition probabilities.}

\item{covs}{Covariate values to include in the model, as a dataframe. The number of rows of \code{covs}
needs to be a multiple of the number of animals, and the same number of observations will be
simulated for each animal. Default : \code{NULL}. Covariates can also be simulated according to a standard
normal distribution, by setting \code{covs} to \code{NULL}, and specifying \code{nbCovs}.}

\item{nbCovs}{Number of covariates to simulate (0 by default). Does not need to be specified of
\code{covs} is specified.}

\item{zeroInflation}{\code{TRUE} if the step length distribution is inflated in zero.
Default : \code{FALSE}. If \code{TRUE}, values for the zero-mass parameters should be
included in \code{stepPar}.}

\item{obsPerAnimal}{Either the number of the number of observations per animal (if single value),
or the bounds of the number of observations per animal (if vector of two values). In the latter case,
the numbers of obervations generated for each animal are uniformously picked from this interval.
Default : \code{c(500,1500)}. \code{obsPerAnimal} does not need to be specified if \code{covs} is
specified.}
}
\value{
An object moveData, i.e. a dataframe of :
\item{ID}{The ID(s) of the observed animal(s)}
\item{step}{The step lengths}
\item{angle}{The turning angles (if any)}
\item{x}{Either easting or longitude}
\item{y}{Either norting or latitude}
\item{...}{Covariates (if any)}
}
\description{
Simulates movement data from an HMM.
}
\examples{
stepPar <- c(1,10,1,5,0.2,0.3) # mean1, mean2, sd1, sd2, z1, z2
anglePar <- c(0,pi,0.5,2) # mean1, mean2, k1, k2
stepDist <- "gamma"
angleDist <- "vm"
obsPerAnimal=c(100,150)
data <- simData(5,2,stepDist,angleDist,stepPar,anglePar,nbCovs=2,zeroInflation=TRUE,
               obsPerAnimal=obsPerAnimal)

stepPar <- c(1,10,1,5) # mean1, mean2, sd1, sd2
anglePar <- c(0,pi,0.5,0.7) # mean1, mean2, k1, k2
stepDist <- "weibull"
angleDist <- "wrpcauchy"
data <- simData(5,2,stepDist,angleDist,stepPar,anglePar,obsPerAnimal=obsPerAnimal)

# step length only and zero-inflation
stepPar <- c(1,10,1,5,0.2,0.3) # mean1, mean2, sd1, sd2, z1, z2
stepDist <- "gamma"
data <- simData(5,2,stepDist,"none",stepPar,nbCovs=2,zeroInflation=TRUE,obsPerAnimal=obsPerAnimal)
}

