% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{Simulation tool}
\usage{
simData(nbAnimals, nbStates, stepDist = c("gamma", "weibull", "lnorm", "exp"),
  angleDist = c("vm", "wrpcauchy", "none"), stepPar, anglePar = NULL,
  beta = NULL, nbCovs = 0, zeroInflation = FALSE, obsPerAnimal = c(500,
  1500))
}
\arguments{
\item{nbAnimals}{Number of observed individuals to simulate.}

\item{nbStates}{Number of behavioural states to simulate.}

\item{stepDist}{Name of the distribution of the step lengths (as a character string).
Supported distributions are : gamma, weibull, lnorm, exp. Default : gamma.}

\item{angleDist}{Name of the distribution of the turning angles (as a character string).
Supported distributions are : vm, wrpcauchy. Set to "none" if the angle distribution should
not be estimated. Default : vm.}

\item{stepPar}{Parameters of the step length distribution.}

\item{anglePar}{Parameters of the turning angle distribution.}

\item{beta}{Matrix of regression parameters for the transition probabilities.}

\item{nbCovs}{Number of covariates to simulate (0 by default).}

\item{zeroInflation}{TRUE if the step length distribution is inflated in zero. Default : FALSE. If TRUE,
values for the zero-mass parameters should be included in stepPar.}

\item{obsPerAnimal}{Bounds of the number of observations per animal. Default : (500,1500).}
}
\value{
An object moveData
}
\description{
Simulates movement data from an HMM.
}
\examples{
stepPar <- c(1,10,1,5,0.2,0.3) # mean1, mean2, sd1, sd2, z1, z2
anglePar <- c(0,pi,0.5,2) # mean1, mean2, k1, k2
stepDist <- "gamma"
angleDist <- "vm"
obsPerAnimal=c(100,150)
data <- simData(5,2,stepDist,angleDist,stepPar,anglePar,nbCovs=2,zeroInflation=TRUE,
               obsPerAnimal=obsPerAnimal)

stepPar <- c(1,10,1,5) # mean1, mean2, sd1, sd2
anglePar <- c(0,pi,0.5,0.7) # mean1, mean2, k1, k2
stepDist <- "weibull"
angleDist <- "wrpcauchy"
data <- simData(5,2,stepDist,angleDist,stepPar,anglePar,obsPerAnimal=obsPerAnimal)

# step length only and zero-inflation
stepPar <- c(1,10,1,5,0.2,0.3) # mean1, mean2, sd1, sd2, z1, z2
stepDist <- "gamma"
data <- simData(5,2,stepDist,"none",stepPar,nbCovs=2,zeroInflation=TRUE,obsPerAnimal=obsPerAnimal)
}

