% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{Simulation tool}
\usage{
simData(nbAnimals, nbStates, stepPar, anglePar, nbCov = 0)
}
\arguments{
\item{nbAnimals}{Number of observed individuals to simulate.}

\item{nbStates}{Number of behavioural states to simulate.}

\item{stepPar}{Parameters of the step length gamma distribution. Must be provided in a
matrix of dimensions 2*nbStates. First row : mean ; second row : standard deviation.}

\item{anglePar}{Parameters of the turning angle von Mises distribution. Must be provided
in a matrix of dimensions 2*nbStates. First row : mean ; second row : concentration power.}

\item{nbCov}{Number of covariates to simulate (0 by default).}
}
\value{
An object moveData
}
\description{
Simulation tool
}
\examples{
stepPar <- matrix(c(1,1,10,5),nrow=2) # mean1, sd1, mean2, sd2
anglePar <- matrix(c(0,0.5,pi,2),nrow=2) # mean1, k1, mean2, k2
data <- simData(5,2,1,stepPar,anglePar)
}

