% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stateProbs.R
\name{stateProbs.moveHMM}
\alias{stateProbs.moveHMM}
\title{State probabilities}
\usage{
\method{stateProbs}{moveHMM}(m)
}
\arguments{
\item{m}{A \code{moveHMM} object.}
}
\value{
The matrix of state probabilities.
}
\description{
State probabilities
}
\examples{
m <- ex$m # moveHMM object (returned by fitHMM)

sp <- stateProbs(m)
}
\references{
Zucchini, W. and MacDonald, I.L. 2009.
Hidden Markov Models for Time Series: An Introduction Using R.
Chapman & Hall (London).
}

