% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/viterbi.R
\name{viterbi}
\alias{viterbi}
\title{Viterbi algorithm}
\usage{
viterbi(data, nbStates, beta, delta, stepDist = c("gamma", "weibull", "exp"),
  angleDist = c("NULL", "vm", "wrpcauchy"), stepPar, anglePar = NULL,
  angleMean = NULL, zeroInflation = FALSE)
}
\arguments{
\item{data}{An object moveData.}

\item{nbStates}{Number of states of the HMM.}

\item{beta}{Matrix of regression coefficients for the transition probability matrix.}

\item{delta}{Stationary distribution.}

\item{stepDist}{Name of the distribution of the step length values.}

\item{angleDist}{Name of the distribution of the turning angle values. Defaults to "NULL"
if the turning angles distributions is not estimated.}

\item{stepPar}{Vector of state-dependent step length distribution parameters.}

\item{anglePar}{Vector of state-dependent turning angle distribution parameters.}

\item{angleMean}{Vector of means of turning angles if not estimated (one for each state).
Defaults to NULL.}

\item{zeroInflation}{TRUE if the step length distribution is inflated in zero.}
}
\value{
The sequence of most probable states.
}
\description{
Reconstructs the most probable states sequence.
}
\examples{
data <- example$data
mod <- example$mod

# reconstruction of states sequence
states <- viterbi(data,nbStates,mod$mle$beta,mod$mle$delta,stepDist,angleDist,mod$mle$stepPar,
                  mod$mle$anglePar,angleMean)
}

